+++ src/main/java/com/nongxinle/controller/NxDistributerGoodsShelfController.java
@@ -11,15 +11,13 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import com.nongxinle.entity.GbDistributerGoodsShelfEntity;
-import com.nongxinle.entity.NxDistributerFatherGoodsEntity;
-import com.nongxinle.entity.NxDistributerGoodsShelfGoodsEntity;
+import com.nongxinle.entity.*;
 import com.nongxinle.service.NxDistributerGoodsShelfGoodsService;
+import com.nongxinle.service.NxDistributerService;
 import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
-import com.nongxinle.entity.NxDistributerGoodsShelfEntity;
 import com.nongxinle.service.NxDistributerGoodsShelfService;
 import com.nongxinle.utils.PageUtils;
 import com.nongxinle.utils.R;
@@ -32,6 +30,8 @@ public class NxDistributerGoodsShelfController {
 	private NxDistributerGoodsShelfService nxDisGoodsShelfService;
 	@Autowired
 	private NxDistributerGoodsShelfGoodsService nxDisGoodsShelfGoodsService;
+	@Autowired
+	private NxDistributerService nxDistributerService;
 
 
 
@@ -41,21 +41,44 @@ public class NxDistributerGoodsShelfController {
 	public R disGetToPlanPurchaseShelfGoods(Integer disId, Integer shelfId) {
 		Map<String, Object> map = new HashMap<>();
 		map.put("disId", disId);
-		map.put("purStatus", 1);
+		map.put("purStatus", 2);
 		map.put("shelfId", shelfId);
 		List<NxDistributerFatherGoodsEntity> shelfGoodsEntities = nxDisGoodsShelfService.disGetUnPlanShelfPurchaseApplys(map);
 
 		return R.ok().put("data", shelfGoodsEntities);
 	}
 
+	@RequestMapping(value = "/updateShelfName", method = RequestMethod.POST)
+	@ResponseBody
+	public R updateShelfName (Integer shelfId, String shelfName) {
+
+		NxDistributerGoodsShelfEntity shelfEntity = nxDisGoodsShelfService.queryObject(shelfId);
+		shelfEntity.setNxDistributerGoodsShelfName(shelfName);
+		nxDisGoodsShelfService.update(shelfEntity);
+		return R.ok().put("data", shelfEntity);
+	}
+
+	@RequestMapping(value = "/updateShelfSort", method = RequestMethod.POST)
+	@ResponseBody
+	public R updateShelfSort (@RequestBody List<NxDistributerGoodsShelfEntity> shelfList) {
+
+		for (int i = 0; i < shelfList.size(); i++){
+			NxDistributerGoodsShelfEntity shelfEntity = shelfList.get(i);
+			shelfEntity.setNxDistributerGoodsShelfSort(i + 1);
+			nxDisGoodsShelfService.update(shelfEntity);
+		}
+		return R.ok();
+	}
 
 	@RequestMapping(value = "/getShelfGoods/{shelfId}")
 	@ResponseBody
 	public R getShelfGoods(@PathVariable Integer shelfId) {
 		Map<String, Object> map = new HashMap<>();
 		map.put("shelfId", shelfId);
-		NxDistributerGoodsShelfEntity shelfEntity =  nxDisGoodsShelfService.queryShelfGoodsByParams(map);
-		return R.ok().put("data",shelfEntity);
+//		NxDistributerGoodsShelfEntity shelfEntity =  nxDisGoodsShelfService.queryShelfGoodsByParams(map);
+		List<NxDistributerGoodsShelfGoodsEntity> nxDistributerGoodsShelfGoodsEntities = nxDisGoodsShelfGoodsService.queryShelfForGoodsByParams(map);
+//		List<NxDistributerGoodsShelfGoodsEntity> nxDistributerGoodsShelfGoodsEntities = nxDisGoodsShelfGoodsService.queryShelfForGoodsWithOrders(map);
+		return R.ok().put("data",nxDistributerGoodsShelfGoodsEntities);
 	}
 
 	@RequestMapping(value = "/disGetShelfs/{disId}")
@@ -67,14 +90,15 @@ public class NxDistributerGoodsShelfController {
 	    return R.ok().put("data", shelfEntities);
 	}
 
-	@RequestMapping(value = "/disGetShelfsWithDetail/{disId}")
-	@ResponseBody
-	public R disGetShelfsWithDetail(@PathVariable Integer disId) {
-		Map<String, Object> map = new HashMap<>();
-		map.put("disId", disId);
-		List<NxDistributerGoodsShelfEntity> shelfEntities = nxDisGoodsShelfService.queryShelfWithDetailByParams(map);
-		return R.ok().put("data", shelfEntities);
-	}
+//
+//	@RequestMapping(value = "/disGetShelfsWithDetail/{disId}")
+//	@ResponseBody
+//	public R disGetShelfsWithDetail(@PathVariable Integer disId) {
+//		Map<String, Object> map = new HashMap<>();
+//		map.put("disId", disId);
+//		List<NxDistributerGoodsShelfEntity> shelfEntities = nxDisGoodsShelfService.queryShelfWithDetailByParams(map);
+//		return R.ok().put("data", shelfEntities);
+//	}
 
 	@RequestMapping(value = "/updateShelf", method = RequestMethod.POST)
 	@ResponseBody
@@ -86,6 +110,8 @@ public class NxDistributerGoodsShelfController {
 //		}
 		nxDisGoodsShelfService.update(shelfEntity);
 
+
+
 		return R.ok();
 	}
 
@@ -93,13 +119,11 @@ public class NxDistributerGoodsShelfController {
 	@ResponseBody
 	public R saveNewShelf (@RequestBody NxDistributerGoodsShelfEntity shelf) {
 	    nxDisGoodsShelfService.save(shelf);
-//		List<NxDistributerGoodsShelfGoodsEntity> nxDisGoodsShelfGoodsEntities = shelf.getNxDisGoodsShelfGoodsEntities();
-//
-//		for (NxDistributerGoodsShelfGoodsEntity shelfGoods : nxDisGoodsShelfGoodsEntities) {
-//			shelfGoods.setNxDgsgShelfId(shelf.getNxDistributerGoodsShelfId());
-//			nxDisGoodsShelfGoodsService.save(shelfGoods);
-//		}
-		return R.ok();
+		Integer nxDistributerGoodsShelfDisId = shelf.getNxDistributerGoodsShelfDisId();
+		NxDistributerEntity nxDistributerEntity = nxDistributerService.queryObject(nxDistributerGoodsShelfDisId);
+		nxDistributerEntity.setNxDistributerShelfQuantity(nxDistributerEntity.getNxDistributerShelfQuantity() + 1);
+		nxDistributerService.update(nxDistributerEntity);
+		return R.ok().put("data", shelf);
 	}
 
 	@RequestMapping(value = "/deleteShelf/{shelfId}")
@@ -107,13 +131,22 @@ public class NxDistributerGoodsShelfController {
 	public R deleteShelf(@PathVariable Integer shelfId) {
 		Map<String, Object> map = new HashMap<>();
 		map.put("shelfId", shelfId);
-		List<NxDistributerGoodsShelfGoodsEntity> shelfGoodsEntities =  nxDisGoodsShelfGoodsService.queryShelfGoodsByParams(map);
 
-		for (NxDistributerGoodsShelfGoodsEntity shelfGoods : shelfGoodsEntities) {
-			nxDisGoodsShelfGoodsService.delete(shelfGoods.getNxDistributerGoodsShelfGoodsId());
+		List<NxDistributerGoodsShelfGoodsEntity> shelfGoodsEntities =  nxDisGoodsShelfGoodsService.queryShelfForGoodsByParams(map);
+		if(shelfGoodsEntities.size() > 0){
+			return R.error(-1,"先删除商品");
+		}else{
+			NxDistributerGoodsShelfEntity shelfEntity = nxDisGoodsShelfService.queryObject(shelfId);
+			Integer nxDistributerGoodsShelfDisId = shelfEntity.getNxDistributerGoodsShelfDisId();
+			NxDistributerEntity nxDistributerEntity = nxDistributerService.queryObject(nxDistributerGoodsShelfDisId);
+			nxDistributerEntity.setNxDistributerShelfQuantity(nxDistributerEntity.getNxDistributerShelfQuantity() - 1);
+			nxDistributerService.update(nxDistributerEntity);
+
+			nxDisGoodsShelfService.delete(shelfId);
+
+			return R.ok();
 		}
-		nxDisGoodsShelfService.delete(shelfId);
-	    return R.ok();
+
 	}
